Fix Quirks mode and <p> margins
When a xhtml contents is loaded with text/html MIME type, the page might
load into Quirks mode. So it is neccessary to load xhtml pages with
application/xhtml+xml MIME type.

If the page gets loaded in Quirks mode, then majority of the modern CSS
styles might not work.

References -
https://stackoverflow.com/questions/8052178/difference-between-document-documentelement-clientheight-and-document-body-clien
https://en.wikipedia.org/wiki/Quirks_mode